scripts/nhl/requirements.txt
requests==2.32.3
pandas==2.2.2
python-dateutil==2.9.0
supabase==2.6.0
tenacity==9.0.0
